backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
faiss-cpu==1.7.4
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
mlflow==2.13.0
prometheus-client==0.20.0
pydantic==2.7.1
httpx==0.27.0
torch==2.2.2
transformers==4.40.2
huggingface_hub==0.19.4
sentence-transformers==2.7.0
 